+++ examples/api300/synergy/unmanaged_device.rb
@@ -9,22 +9,22 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-require_relative '../_client' # Gives access to @client
+require_relative '../../_client' # Gives access to @client
 
 puts "\nAdding first device ..."
-device1 = OneviewSDK::UnmanagedDevice.new(@client, name: 'Unmanaged Device 1', model: 'Procurve 4200VL')
+device1 = OneviewSDK::API300::Synergy::UnmanagedDevice.new(@client, name: 'Unmanaged Device 1', model: 'Procurve 4200VL')
 device1.add
 puts 'Device added:'
 puts "- \"#{device1['name']}\" with model \"#{device1['model']}\" and uri \"#{device1['uri']}\""
 
 puts "\nAdding second device ..."
-device2 = OneviewSDK::UnmanagedDevice.new(@client, name: 'Unmanaged Device 2', model: 'Unknown')
+device2 = OneviewSDK::API300::Synergy::UnmanagedDevice.new(@client, name: 'Unmanaged Device 2', model: 'Unknown')
 device2.add
 puts 'Device added:'
 puts "- \"#{device2['name']}\" with model \"#{device2['model']}\" and uri \"#{device2['uri']}\""
 
 puts "\nListing all devices:"
-unmanaged_devices = OneviewSDK::UnmanagedDevice.get_devices(@client)
+unmanaged_devices = OneviewSDK::API300::Synergy::UnmanagedDevice.get_devices(@client)
 unmanaged_devices.each do |device|
   puts "- \"#{device['name']}\" with model \"#{device['model']}\" and uri \"#{device['uri']}\""
 end